apps/gateway/pyproject.toml
[project]
name = "ironclaw-gateway"
version = "0.1.0"
description = "Iron Claw Gateway - FastAPI backend for mobile automation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "droidrun>=0.4.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.28.0",
    "aiofiles>=24.0.0",
    "pypdf2>=3.0.0",
    "timezonefinder>=6.5.0",
    "python-telegram-bot>=21.0",
    "llama-index-llms-google-genai>=0.1.0",
    "pyyaml>=6.0.0",
    # Workspace packages
    "ironclaw-shared",
    "ironclaw-adb",
    "ironclaw-vapi",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ironclaw"]
